rename distiller-pipeline -> distiller-streaming
