+++ operators/distiller-streaming/distiller_streaming/emitter.py
@@ -3,7 +3,7 @@ from collections.abc import Generator
 import numpy as np
 from stempy.io import SparseArray
 
-from distiller_pipeline.models import FrameHeader
+from distiller_streaming.models import FrameHeader
 from interactem.core.logger import get_logger
 from interactem.core.models.messages import BytesMessage, MessageHeader, MessageSubject
